Changed type of field Year Modern.
git-svn-id: https://svn.java.net/svn/docsources~svn/trunk@60 3b25252b-017b-4ae5-8f61-fa87a855a72e

+++ DocSources/src/main/java/org/medici/docsources/domain/Document.java
@@ -115,7 +115,7 @@ public class Document implements Serializable{
 	@Column (name="\"SORTABLEDATE\"", length=50)
 	private String sortableDate;
 	@Column (name="\"YEARMODERN\"", length=15, precision=5)
-	private Double yearModern;
+	private Integer yearModern;
 	@Column (name="\"RECKONING\"", length=1, columnDefinition="tinyint", nullable=false)
 	private Boolean reckoning;
 	@Column (name="\"UNDATED\"", length=1, columnDefinition="tinyint", nullable=false)
@@ -440,13 +440,13 @@ public class Document implements Serializable{
 	/**
 	 * @return the yearModern
 	 */
-	public Double getYearModern() {
+	public Integer getYearModern() {
 		return yearModern;
 	}
 	/**
 	 * @param yearModern the yearModern to set
 	 */
-	public void setYearModern(Double yearModern) {
+	public void setYearModern(Integer yearModern) {
 		this.yearModern = yearModern;
 	}
 	/**
